feat: 파일, 클래스 명 변경, 앱바 추가
